+++ template_example/example/standalone/providers.tf
@@ -10,3 +10,18 @@ provider "nomad" {
   secret_id = var.nomad_acl ? data.vault_generic_secret.nomad_secret_id[0].data.secret_id : null
 }
 
+provider "vault" {
+  address = "http://127.0.0.1:8200"
+  token = "master"
+}
+
+terraform {
+  required_providers {
+    vault = {
+      version = ">=2.13.0"
+    }
+    nomad = {
+      version = ">=1.4.0"
+    }
+  }
+}